+++ react-components/src/architecture/concrete/annotations/commands/AnnotationsAlignCommand.ts
