+++ react-components/src/architecture/concrete/annotations/commands/AnnotationsShowCommand.ts
@@ -2,20 +2,20 @@
  * Copyright 2024 Cognite AS
  */
 
-import { ShowDomainObjectsOnTopCommand } from '../../../base/commands/ShowDomainObjectsOnTopCommand';
+import { ShowAllDomainObjectsCommand } from '../../../base/commands/ShowAllDomainObjectsCommand';
 import { type DomainObject } from '../../../base/domainObjects/DomainObject';
 import { type TranslateKey } from '../../../base/utilities/TranslateKey';
 import { AnnotationsDomainObject } from '../AnnotationsDomainObject';
 import { BoxGizmoDomainObject } from '../BoxGizmoDomainObject';
 import { CylinderGizmoDomainObject } from '../CylinderGizmoDomainObject';
 
-export class ShowAnnotationsOnTopCommand extends ShowDomainObjectsOnTopCommand {
+export class AnnotationsShowCommand extends ShowAllDomainObjectsCommand {
   // ==================================================
   // OVERRIDES
   // ==================================================
 
   public override get tooltip(): TranslateKey {
-    return { key: 'ANNOTATIONS_SHOW_ON_TOP', fallback: 'Show all annotations on top' };
+    return { key: 'ANNOTATIONS_SHOW_OR_HIDE', fallback: 'Show or hide annotations' };
   }
 
   protected override isInstance(domainObject: DomainObject): boolean {
