+++ react-components/src/architecture/concrete/annotations/commands/AnnotationsShowOnTopCommand.ts
@@ -2,20 +2,20 @@
  * Copyright 2024 Cognite AS
  */
 
-import { ShowAllDomainObjectsCommand } from '../../../base/commands/ShowAllDomainObjectsCommand';
+import { ShowDomainObjectsOnTopCommand } from '../../../base/commands/ShowDomainObjectsOnTopCommand';
 import { type DomainObject } from '../../../base/domainObjects/DomainObject';
 import { type TranslateKey } from '../../../base/utilities/TranslateKey';
 import { AnnotationsDomainObject } from '../AnnotationsDomainObject';
 import { BoxGizmoDomainObject } from '../BoxGizmoDomainObject';
 import { CylinderGizmoDomainObject } from '../CylinderGizmoDomainObject';
 
-export class ShowAllAnnotationsCommand extends ShowAllDomainObjectsCommand {
+export class AnnotationsShowOnTopCommand extends ShowDomainObjectsOnTopCommand {
   // ==================================================
   // OVERRIDES
   // ==================================================
 
   public override get tooltip(): TranslateKey {
-    return { key: 'ANNOTATIONS_SHOW_OR_HIDE', fallback: 'Show or hide annotations' };
+    return { key: 'ANNOTATIONS_SHOW_ON_TOP', fallback: 'Show all annotations on top' };
   }
 
   protected override isInstance(domainObject: DomainObject): boolean {
