+++ react-components/src/architecture/concrete/config/StoryBookConfig.ts
@@ -24,8 +24,11 @@ import { ObservationsTool } from '../observations/ObservationsTool';
 import { SettingsCommand } from '../../base/concreteCommands/SettingsCommand';
 import { MockSettingsCommand } from '../../base/commands/mocks/MockSettingsCommand';
 import { MockFilterCommand } from '../../base/commands/mocks/MockFilterCommand';
-import { AnnotationEditTool } from '../annotations/commands/AnnotationEditTool';
+import { AnnotationsCreateTool } from '../annotations/commands/AnnotationsCreateTool';
 import { ToggleAllModelsVisibleCommand } from '../../base/concreteCommands/ToggleAllModelsVisibleCommand';
+import { AnnotationsShowCommand } from '../annotations/commands/AnnotationsShowCommand';
+import { AnnotationsShowOnTopCommand } from '../annotations/commands/AnnotationsShowOnTopCommand';
+import { AnnotationsSelectTool } from '../annotations/commands/AnnotationsSelectTool';
 
 export class StoryBookConfig extends BaseRevealConfig {
   // ==================================================
@@ -50,12 +53,17 @@ export class StoryBookConfig extends BaseRevealConfig {
       new MockSettingsCommand(),
       new MockFilterCommand(),
       undefined,
+      new AnnotationsSelectTool(),
+      new AnnotationsCreateTool(),
+      new AnnotationsShowCommand(),
+      new AnnotationsShowOnTopCommand(),
+      undefined,
+      new ObservationsTool(),
+      undefined,
       new ExampleTool(),
       new MeasurementTool(),
       new ClipTool(),
       new ObservationsTool(),
-      undefined,
-      new AnnotationEditTool(),
       undefined,
       new SetTerrainVisibleCommand(),
       new UpdateTerrainCommand(),
